refactor: spilt map into lots of components
